+++ lets/api/rest/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path("<path:module>", views.do, name="do")
+]