+++ lets/modules/analyze/disassemble/x86.py
@@ -7,6 +7,7 @@ class X86(DockerExtension, Module):
     """
     Disassemble bytes into x86 assembly with an interactive radare2 console.
     """
+    interactive = True
 
     # A list of docker images required by the module.
     images = [